WIP
I did way to much here without committing, and I'm not sure any specific
mutation is complete. Going to pick it apart and recreate it one step (and
commit) at a time on master, but I wanted it in the repo on a branch for both
history and to be able to transfer it between computers.

+++ src/model/BlackBruise.java
@@ -11,11 +11,11 @@ public class BlackBruise {
 		if (heroic) {
 			// TODO: EXPAND
 			// B1076
-			return (double) 0;
+			return 0;
 		} else {
 			// TODO: EXPAND
 			// B1075
-			return (double) 0;
+			return 0;
 		}
 	}
 
@@ -31,6 +31,6 @@ public class BlackBruise {
 
 	public double magicalDPS(Character character, Target target) {
 		return character.ferociousInspiration() * target.magicSuppression()
-				* multiplier() * uptime(character) * character.rawWhiteDPS();
+				* multiplier() * uptime(character) * character.rawWhiteDPS(target);
 	}
 }
